+++ packages/shared-components/src/StudentQuestion.tsx
@@ -32,7 +32,10 @@ function ImgWithModal({ src, alt }: { src: string; alt?: string }) {
       open={isOpen}
       trigger={
         <Button basic onClick={() => setIsOpen(true)}>
-          <Image src={src} alt="Image" width={250} height={250} />
+          <div className="flex flex-col items-start mb-1">
+            <Image src={src} alt="Image" width={250} height={250} />
+            {alt && <div className="text-sm text-slate-600">{alt}</div>}
+          </div>
         </Button>
       }
       onClose={() => setIsOpen(false)}